Removing lot of warnings
git-svn-id: https://svn.apache.org/repos/asf/directory/trunks/shared@487348 13f79535-47bb-0310-9956-ffa450edef68

+++ ldap/src/main/java/org/apache/directory/shared/ldap/codec/add/AddRequest.java
@@ -372,7 +372,7 @@ public class AddRequest extends LdapMessage
 
                     // The attributes list sequence
                     buffer.put( UniversalTag.SEQUENCE_TAG );
-                    int localAttributeLength = ( ( Integer ) attributeLength.get( attributeNumber ) ).intValue();
+                    int localAttributeLength = attributeLength.get( attributeNumber );
                     buffer.put( TLV.getBytes( localAttributeLength ) );
 
                     // The attribute type
@@ -380,7 +380,7 @@ public class AddRequest extends LdapMessage
 
                     // The values
                     buffer.put( UniversalTag.SET_TAG );
-                    int localValuesLength = ( ( Integer ) valuesLength.get( attributeNumber ) ).intValue();
+                    int localValuesLength = valuesLength.get( attributeNumber );
                     buffer.put( TLV.getBytes( localValuesLength ) );
 
                     try
